requirements.txt
django djangorestframework
Pillow
djangorestframework-simplejwt
django-ratelimit
django-ratelimit
drf-yasg
